refactor: Add reusable components for later use
In this commit, I have added 4 main components that are going to be used
later. Those components are ( Header, Card, CardSection and Button). They
are going to be useful in creating the "auth" project since we are going
to create a login form that needs the components that we added.