feat: Show proper form on the home screen
When a user opens the app, it will show to him a Spinner
while we check if they are logged in or not, if they are,
the app will show a "Log Out". Otherwise, the app will
show to them "Login Form".

Note: this commit contains a buggy button when you login. You need
to revisit this commit.